// types/zod/inputTypeSchemas/driver_municipalitiesMaxOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const driver_municipalitiesMaxOrderByAggregateInputSchema: z.ZodType<Prisma.driver_municipalitiesMaxOrderByAggregateInput> = z.object({
  driver_municipalities_id: z.lazy(() => SortOrderSchema).optional(),
  driver_id: z.lazy(() => SortOrderSchema).optional(),
  municipalities_id: z.lazy(() => SortOrderSchema).optional(),
  created_at: z.lazy(() => SortOrderSchema).optional(),
  updated_at: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default driver_municipalitiesMaxOrderByAggregateInputSchema;